+++ ng/practices-ng-list-view-source/src/lib/select-view-source-fluent-builder/types.ts
@@ -1,16 +1,45 @@
 import { Type } from '@angular/core';
 import { MaybeAsync } from '@angular/router';
 import { StringKeyOf } from '@nexplore/practices-ng-common-util';
-import { SelectViewSource, SelectViewSourceConfig } from '../implementation/select-view-source';
+import { SelectViewSource } from '../implementation/select-view-source';
 import { EnhancedListViewSource, IListResult, IQueryParamsWithFilter } from '../types';
 import { HasTypedQueryParams } from '../types-internal';
 
+export type SelectViewSourceDefaultFilter = { label: string };
+
+/**
+ * The default key used by ng-select, when there is not bind-label used
+ */
+export type SelectViewSourceDefaultLabelKey = keyof SelectViewSourceDefaultFilter;
+
+export type SelectViewSourceLabelKey<TData> = NoInfer<StringKeyOf<TData>> | SelectViewSourceDefaultLabelKey;
+export type SelectViewSourceFilterType<TData, TLabelKey> = [TLabelKey] extends [StringKeyOf<TData>]
+    ? Pick<TData, TLabelKey>
+    : SelectViewSourceDefaultFilter;
+
 export type TypedSelectViewSourceConfig<
     TData,
-    TLabelKey extends NoInfer<StringKeyOf<TData>>,
+    TLabelKey extends SelectViewSourceLabelKey<TData>,
     TQueryParams = TData,
-    TOrdering = Pick<TData, TLabelKey>
-> = SelectViewSourceConfig<TData> & {
+    TOrdering = TData
+> = (TLabelKey extends StringKeyOf<TData>
+    ? {
+          // The label property to display in the select
+          label?: TLabelKey;
+      }
+    : {
+          // The label property to display in the select
+          label?: StringKeyOf<TData>;
+      }) & {
+    // The value property to bind
+    value?: keyof TData;
+
+    // True when the select should be searchable, false otherwise. Default is true.
+    searchable?: boolean;
+
+    // True when the select should filter interally using a simple text search, false otherwise. Default is false.
+    localSearch?: boolean;
+
     /** Type is used solely for type inference. Unfortunately, typescript cannot properly infer the type without it */
     type?: Type<TData>;
 
@@ -22,7 +51,9 @@ export type TypedSelectViewSourceConfig<
      * */
     queryParamsType?: Type<TQueryParams>;
 
-    loadFn: (params: IQueryParamsWithFilter<Pick<TData, TLabelKey>>) => MaybeAsync<Partial<IListResult<TData>>>;
+    loadFn: (
+        params: IQueryParamsWithFilter<SelectViewSourceFilterType<TData, TLabelKey>>
+    ) => MaybeAsync<Partial<IListResult<TData>>>;
 } & HasTypedQueryParams<TQueryParams, TOrdering>;
 
 export type SelectViewSourceWithSignals<TData, TFilter> = EnhancedListViewSource<SelectViewSource<TData, TFilter>>;